small changes and bugfixes
